--- src/app/_component/section/banner/DynamicBanner.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-const DynamicBanner = dynamic(() => import('./ActualBanner.tsx'), {
-  ssr: false,
-});
-
-export default DynamicBanner;
