--- src/app/_component/section/banner/index.ts
@@ -1,1 +1,0 @@
-export { default as Banner } from './DynamicBanner.tsx';
