--- src/app/_component/section/carousel/DynamicCarouselBanner.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-const DynamicCarouselBanner = dynamic(() => import('./ActualCarouselBanner.tsx'), {
-  ssr: false,
-});
-
-export default DynamicCarouselBanner;
