--- src/app/_component/section/carousel/index.ts
@@ -1,1 +1,0 @@
-export { default as CarouselBanner } from './DynamicCarouselBanner.tsx';
