--- src/app/_component/section/description/DynamicDescription.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-const DynamicDescription = dynamic(() => import('./ActualDescription.tsx'), {
-  ssr: false,
-});
-
-export default DynamicDescription;
