--- src/app/_component/section/description/index.ts
@@ -1,1 +1,0 @@
-export { default as Description } from './DynamicDescription.tsx';
