--- src/app/_component/section/feature/DynamicFeature.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-const DynamicFeature = dynamic(() => import('./ActualFeature.tsx'), {
-  ssr: false,
-});
-
-export default DynamicFeature;
