--- src/app/_component/section/feature/DynamicFeatureItem.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-const DynamicFeatureItem = dynamic(() => import('./ActualFeatureItem.tsx'), {
-  ssr: false,
-});
-
-export default DynamicFeatureItem;
