--- src/app/_component/section/feature/index.ts
@@ -1,2 +1,0 @@
-export { default as Feature } from './DynamicFeature.tsx';
-export { default as FeatureItem } from './DynamicFeatureItem.tsx';
