+++ src/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useIsMobile } from './useIsMobile';
+export { useIsTablet } from './useIsTablet';
